requirements.txt
discord.py==1.7.3
heroku==0.1.4
pip==21.1.3
psycopg2==2.9.1
PyNaCl==1.4.0